requirements.txt
freetype-py

numpy-stl

PyOpenGL
OpenGLContext
PyVRML97
pydispatcher
numpy
pillow

